+++ protocol/common/message.go
@@ -8,7 +8,6 @@ import (
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"net/url"
 	"os"
 	"strings"
 	"unicode"
@@ -81,60 +80,6 @@ const (
 	ContactVerificationStateUntrustworthy
 	ContactVerificationStateCanceled
 )
-
-type LinkPreviewThumbnail struct {
-	Width  int `json:"width"`
-	Height int `json:"height"`
-	// Non-empty when the thumbnail is available via the media server, i.e. after
-	// the chat message is sent.
-	URL string `json:"url,omitempty"`
-	// Non-empty when the thumbnail payload needs to be shared with the client,
-	// but before it has been persisted.
-	DataURI string `json:"dataUri,omitempty"`
-}
-
-type LinkPreview struct {
-	Type        protobuf.UnfurledLink_LinkType `json:"type"`
-	URL         string                         `json:"url"`
-	Hostname    string                         `json:"hostname"`
-	Title       string                         `json:"title,omitempty"`
-	Description string                         `json:"description,omitempty"`
-	Thumbnail   LinkPreviewThumbnail           `json:"thumbnail,omitempty"`
-}
-
-type StatusContactLinkPreview struct {
-	PublicKey   string               `json:"publicKey"`
-	DisplayName string               `json:"displayName"`
-	Description string               `json:"description"`
-	Icon        LinkPreviewThumbnail `json:"icon,omitempty"`
-}
-
-type StatusCommunityLinkPreview struct {
-	CommunityID  string               `json:"communityId"`
-	DisplayName  string               `json:"displayName"`
-	Description  string               `json:"description"`
-	MembersCount uint32               `json:"membersCount"`
-	Color        string               `json:"color"`
-	TagIndices   []uint32             `json:"tagIndices"`
-	Icon         LinkPreviewThumbnail `json:"icon,omitempty"`
-	Banner       LinkPreviewThumbnail `json:"banner,omitempty"`
-}
-
-type StatusCommunityChannelLinkPreview struct {
-	ChannelUUID string                      `json:"channelUuid"`
-	Emoji       string                      `json:"emoji"`
-	DisplayName string                      `json:"displayName"`
-	Description string                      `json:"description"`
-	Color       string                      `json:"color"`
-	Community   *StatusCommunityLinkPreview `json:"community"`
-}
-
-type StatusLinkPreview struct {
-	URL       string                             `json:"url,omitempty"`
-	Contact   *StatusContactLinkPreview          `json:"contact,omitempty"`
-	Community *StatusCommunityLinkPreview        `json:"community,omitempty"`
-	Channel   *StatusCommunityChannelLinkPreview `json:"channel,omitempty"`
-}
 
 const EveryoneMentionTag = "0x00001"
 
@@ -798,392 +743,6 @@ func (m *Message) LoadImage() error {
 	m.Payload = &protobuf.ChatMessage_Image{Image: imageMessage}
 
 	return nil
-}
-
-func isValidLinkPreviewThumbnail(thumbnail LinkPreviewThumbnail) error {
-	if thumbnail.DataURI == "" {
-		if thumbnail.Width == 0 && thumbnail.Height == 0 {
-			return nil
-		}
-		return fmt.Errorf("dataUri is empty, but width/height are not zero")
-	}
-
-	if thumbnail.Width == 0 || thumbnail.Height == 0 {
-		return fmt.Errorf("dataUri is not empty, but width/heigth are zero")
-	}
-
-	return nil
-}
-
-func isValidLinkPreviewForProto(preview LinkPreview) error {
-	switch preview.Type {
-	case protobuf.UnfurledLink_IMAGE:
-		if preview.URL == "" {
-			return fmt.Errorf("empty url")
-		}
-		if err := isValidLinkPreviewThumbnail(preview.Thumbnail); err != nil {
-			return fmt.Errorf("thumbnail is not valid for proto: %w", err)
-		}
-		return nil
-	default: // Validate as a link type by default.
-		if preview.Title == "" {
-			return fmt.Errorf("title is empty")
-		}
-		if preview.URL == "" {
-			return fmt.Errorf("url is empty")
-		}
-		if err := isValidLinkPreviewThumbnail(preview.Thumbnail); err != nil {
-			return fmt.Errorf("thumbnail is not valid for proto: %w", err)
-		}
-		return nil
-	}
-}
-
-func (preview *StatusCommunityLinkPreview) isValidForProto() error {
-	if preview == nil {
-		return fmt.Errorf("community preview is empty")
-	}
-	if preview.CommunityID == "" {
-		return fmt.Errorf("communityId is empty")
-	}
-	if err := isValidLinkPreviewThumbnail(preview.Icon); err != nil {
-		return fmt.Errorf("community icon is invalid: %w", err)
-	}
-	if err := isValidLinkPreviewThumbnail(preview.Banner); err != nil {
-		return fmt.Errorf("community banner is invalid: %w", err)
-	}
-	return nil
-}
-
-func (preview StatusLinkPreview) isValidForProto() error {
-	if preview.URL == "" {
-		return fmt.Errorf("url can't be empty")
-	}
-
-	// At least and only one of Contact/Community/Channel can be present in the preview
-	if preview.Contact != nil && preview.Community != nil ||
-		preview.Community != nil && preview.Channel != nil ||
-		preview.Channel != nil && preview.Contact != nil {
-		return fmt.Errorf("several of contact/community/channel are set at the same time")
-	}
-
-	if preview.Contact == nil && preview.Community == nil && preview.Channel == nil {
-		return fmt.Errorf("none of contact/community/channel are set")
-	}
-
-	if preview.Contact != nil {
-		if preview.Contact.PublicKey == "" {
-			return fmt.Errorf("contact publicKey is empty")
-		}
-		if err := isValidLinkPreviewThumbnail(preview.Contact.Icon); err != nil {
-			return fmt.Errorf("contact icon invalid: %w", err)
-		}
-		return nil
-	}
-
-	if preview.Community != nil {
-		return preview.Community.isValidForProto()
-	}
-
-	if preview.Channel != nil {
-		if preview.Channel.ChannelUUID == "" {
-			return fmt.Errorf("channelUuid is empty")
-		}
-		if preview.Channel.Community == nil {
-			return fmt.Errorf("channel community is nil")
-		}
-		if err := preview.Channel.Community.isValidForProto(); err != nil {
-			return fmt.Errorf("channel community is not valid: %w", err)
-		}
-		return nil
-	}
-	return nil
-}
-
-// ConvertLinkPreviewsToProto expects previews to be correctly sent by the
-// client because we can't attempt to re-unfurl URLs at this point (it's
-// actually undesirable). We run a basic validation as an additional safety net.
-func (m *Message) ConvertLinkPreviewsToProto() ([]*protobuf.UnfurledLink, error) {
-	if len(m.LinkPreviews) == 0 {
-		return nil, nil
-	}
-
-	unfurledLinks := make([]*protobuf.UnfurledLink, 0, len(m.LinkPreviews))
-
-	for _, preview := range m.LinkPreviews {
-		// Do not process subsequent previews because we do expect all previews to
-		// be valid at this stage.
-		if err := isValidLinkPreviewForProto(preview); err != nil {
-			return nil, fmt.Errorf("invalid link preview, url='%s': %w", preview.URL, err)
-		}
-
-		var payload []byte
-		var err error
-		if preview.Thumbnail.DataURI != "" {
-			payload, err = images.GetPayloadFromURI(preview.Thumbnail.DataURI)
-			if err != nil {
-				return nil, fmt.Errorf("could not get data URI payload, url='%s': %w", preview.URL, err)
-			}
-		}
-
-		ul := &protobuf.UnfurledLink{
-			Type:             preview.Type,
-			Url:              preview.URL,
-			Title:            preview.Title,
-			Description:      preview.Description,
-			ThumbnailWidth:   uint32(preview.Thumbnail.Width),
-			ThumbnailHeight:  uint32(preview.Thumbnail.Height),
-			ThumbnailPayload: payload,
-		}
-		unfurledLinks = append(unfurledLinks, ul)
-	}
-
-	return unfurledLinks, nil
-}
-
-func (m *Message) ConvertFromProtoToLinkPreviews(makeMediaServerURL func(msgID string, previewURL string) string) []LinkPreview {
-	var links []*protobuf.UnfurledLink
-
-	if links = m.GetUnfurledLinks(); links == nil {
-		return nil
-	}
-
-	previews := make([]LinkPreview, 0, len(links))
-	for _, link := range links {
-		parsedURL, err := url.Parse(link.Url)
-		var hostname string
-		// URL parsing in Go can fail with URLs that weren't correctly URL encoded.
-		// This shouldn't happen in general, but if an error happens we just reuse
-		// the full URL.
-		if err != nil {
-			hostname = link.Url
-		} else {
-			hostname = parsedURL.Hostname()
-		}
-		lp := LinkPreview{
-			Description: link.Description,
-			Hostname:    hostname,
-			Title:       link.Title,
-			Type:        link.Type,
-			URL:         link.Url,
-		}
-		mediaURL := ""
-		if len(link.ThumbnailPayload) > 0 {
-			mediaURL = makeMediaServerURL(m.ID, link.Url)
-		}
-		if link.GetThumbnailPayload() != nil {
-			lp.Thumbnail.Width = int(link.ThumbnailWidth)
-			lp.Thumbnail.Height = int(link.ThumbnailHeight)
-			lp.Thumbnail.URL = mediaURL
-		}
-		previews = append(previews, lp)
-	}
-
-	return previews
-}
-
-func (preview *LinkPreviewThumbnail) IsEmpty() bool {
-	return preview.Width == 0 &&
-		preview.Height == 0 &&
-		preview.URL == "" &&
-		preview.DataURI == ""
-}
-
-func (preview *LinkPreviewThumbnail) ConvertToProto() (*protobuf.UnfurledLinkThumbnail, error) {
-	var payload []byte
-	var err error
-	if preview.DataURI != "" {
-		payload, err = images.GetPayloadFromURI(preview.DataURI)
-		if err != nil {
-			return nil, fmt.Errorf("could not get data URI payload, url='%s': %w", preview.URL, err)
-		}
-	}
-
-	return &protobuf.UnfurledLinkThumbnail{
-		Width:   uint32(preview.Width),
-		Height:  uint32(preview.Height),
-		Payload: payload,
-	}, nil
-}
-
-func (preview *StatusCommunityLinkPreview) ConvertToProto() (*protobuf.UnfurledStatusCommunityLink, error) {
-	if preview == nil {
-		return nil, nil
-	}
-
-	icon, err := preview.Icon.ConvertToProto()
-	if err != nil {
-		return nil, err
-	}
-
-	banner, err := preview.Banner.ConvertToProto()
-	if err != nil {
-		return nil, err
-	}
-
-	community := &protobuf.UnfurledStatusCommunityLink{
-		CommunityId:  preview.CommunityID,
-		DisplayName:  preview.DisplayName,
-		Description:  preview.Description,
-		MembersCount: preview.MembersCount,
-		Color:        preview.Color,
-		TagIndices:   preview.TagIndices,
-		Icon:         icon,
-		Banner:       banner,
-	}
-
-	return community, nil
-}
-
-func (m *Message) ConvertStatusLinkPreviewsToProto() (*protobuf.UnfurledStatusLinks, error) {
-	if len(m.StatusLinkPreviews) == 0 {
-		return nil, nil
-	}
-
-	unfurledLinks := make([]*protobuf.UnfurledStatusLink, 0, len(m.StatusLinkPreviews))
-
-	for _, preview := range m.StatusLinkPreviews {
-		// We expect all previews to be valid at this stage
-		if err := preview.isValidForProto(); err != nil {
-			return nil, fmt.Errorf("invalid status link preview, url='%s': %w", preview.URL, err)
-		}
-
-		ul := &protobuf.UnfurledStatusLink{
-			Url: preview.URL,
-		}
-
-		if preview.Contact != nil {
-			icon, err := preview.Contact.Icon.ConvertToProto()
-			if err != nil {
-				return nil, err
-			}
-			ul.Payload = &protobuf.UnfurledStatusLink_Contact{
-				Contact: &protobuf.UnfurledStatusContactLink{
-					PublicKey:   preview.Contact.PublicKey,
-					DisplayName: preview.Contact.DisplayName,
-					Description: preview.Contact.Description,
-					Icon:        icon,
-				},
-			}
-		}
-
-		if preview.Community != nil {
-			communityPreview, err := preview.Community.ConvertToProto()
-			if err != nil {
-				return nil, err
-			}
-			ul.Payload = &protobuf.UnfurledStatusLink_Community{
-				Community: communityPreview,
-			}
-		}
-
-		if preview.Channel != nil {
-			communityPreview, err := preview.Channel.Community.ConvertToProto()
-			if err != nil {
-				return nil, err
-			}
-
-			ul.Payload = &protobuf.UnfurledStatusLink_Channel{
-				Channel: &protobuf.UnfurledStatusChannelLink{
-					ChannelUuid: preview.Channel.ChannelUUID,
-					Emoji:       preview.Channel.Emoji,
-					DisplayName: preview.Channel.DisplayName,
-					Description: preview.Channel.Description,
-					Color:       preview.Channel.Color,
-					Community:   communityPreview,
-				},
-			}
-
-		}
-
-		unfurledLinks = append(unfurledLinks, ul)
-	}
-
-	return &protobuf.UnfurledStatusLinks{UnfurledStatusLinks: unfurledLinks}, nil
-}
-
-func (m *Message) ConvertFromProtoToStatusLinkPreviews(makeMediaServerURL func(msgID string, previewURL string, imageId string) string) []StatusLinkPreview {
-
-	if m.GetUnfurledStatusLinks() == nil {
-		return nil
-	}
-
-	links := m.UnfurledStatusLinks.GetUnfurledStatusLinks()
-
-	if links == nil {
-		return nil
-	}
-
-	createThumbnail := func(thumbnail *protobuf.UnfurledLinkThumbnail, URL string, imageId string) LinkPreviewThumbnail {
-		mediaURL := ""
-		if len(thumbnail.Payload) > 0 {
-			mediaURL = makeMediaServerURL(m.ID, URL, imageId)
-		}
-		return LinkPreviewThumbnail{
-			Width:  int(thumbnail.Width),
-			Height: int(thumbnail.Height),
-			URL:    mediaURL,
-		}
-	}
-
-	convertCommunityFromProto := func(c *protobuf.UnfurledStatusCommunityLink, URL string, thumbnailPrefix string) *StatusCommunityLinkPreview {
-		if c == nil {
-			return nil
-		}
-
-		out := &StatusCommunityLinkPreview{
-			CommunityID:  c.CommunityId,
-			DisplayName:  c.DisplayName,
-			Description:  c.Description,
-			MembersCount: c.MembersCount,
-			Color:        c.Color,
-			TagIndices:   c.TagIndices,
-		}
-		if icon := c.GetIcon(); icon != nil {
-			out.Icon = createThumbnail(icon, URL, thumbnailPrefix+"icon")
-		}
-		if banner := c.GetBanner(); banner != nil {
-			out.Banner = createThumbnail(banner, URL, thumbnailPrefix+"banner")
-		}
-		return out
-	}
-
-	previews := make([]StatusLinkPreview, 0, len(links))
-	for _, link := range links {
-		lp := StatusLinkPreview{
-			URL: link.Url,
-		}
-
-		if c := link.GetContact(); c != nil {
-			lp.Contact = &StatusContactLinkPreview{
-				PublicKey:   c.PublicKey,
-				DisplayName: c.DisplayName,
-				Description: c.Description,
-			}
-			if icon := c.GetIcon(); icon != nil {
-				lp.Contact.Icon = createThumbnail(icon, link.Url, "contact-icon")
-			}
-		}
-
-		if c := link.GetCommunity(); c != nil {
-			lp.Community = convertCommunityFromProto(c, link.Url, "community-")
-		}
-
-		if c := link.GetChannel(); c != nil {
-			lp.Channel = &StatusCommunityChannelLinkPreview{
-				ChannelUUID: c.ChannelUuid,
-				Emoji:       c.Emoji,
-				DisplayName: c.DisplayName,
-				Description: c.Description,
-				Color:       c.Color,
-			}
-			lp.Channel.Community = convertCommunityFromProto(c.Community, link.Url, "community-channel-")
-		}
-
-		previews = append(previews, lp)
-	}
-
-	return previews
 }
 
 func (m *Message) SetAlbumIDAndImagesCount(albumID string, imagesCount uint32) error {
