+++ server/handlers_test.go
@@ -3,33 +3,44 @@ package server
 import (
 	"database/sql"
 	"encoding/json"
+	"github.com/status-im/status-go/protocol/tt"
+	"github.com/stretchr/testify/suite"
+	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"testing"
 
-	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 
-	"github.com/status-im/status-go/appdatabase"
-	"github.com/status-im/status-go/logutils"
 	"github.com/status-im/status-go/protocol/common"
 	"github.com/status-im/status-go/protocol/protobuf"
 	"github.com/status-im/status-go/protocol/sqlite"
 	"github.com/status-im/status-go/t/helpers"
 )
 
-func setupTest(t *testing.T) (*sql.DB, *zap.Logger) {
-	db, err := helpers.SetupTestMemorySQLDB(appdatabase.DbInitializer{})
-	require.NoError(t, err)
-	err = sqlite.Migrate(db)
-	require.NoError(t, err)
-
-	logger := logutils.ZapLogger()
-	return db, logger
+func TestHandlersSuite(t *testing.T) {
+	suite.Run(t, new(HandlersSuite))
 }
 
-func createUserMessage(t *testing.T, db *sql.DB, msg *common.Message) {
+type HandlersSuite struct {
+	suite.Suite
+	db     *sql.DB
+	logger *zap.Logger
+}
+
+func (s *HandlersSuite) SetupTest() {
+	dbPath, err := ioutil.TempFile("", "status-go-test-db-")
+	s.Require().NoError(err)
+
+	db, err := sqlite.Open(dbPath.Name(), "", sqlite.ReducedKDFIterationsNumber)
+	s.Require().NoError(err)
+
+	s.logger = tt.MustCreateTestLogger()
+	s.db = db
+}
+
+func (s *HandlersSuite) createUserMessage(msg *common.Message) {
 	whisperTimestamp := 0
 	source := ""
 	text := ""
@@ -40,7 +51,7 @@ func createUserMessage(t *testing.T, db *sql.DB, msg *common.Message) {
 	responseTo := ""
 	clockValue := 0
 
-	stmt, err := db.Prepare(`
+	stmt, err := s.db.Prepare(`
 		INSERT INTO user_messages (
 			id,
 			whisper_timestamp,
@@ -56,13 +67,14 @@ func createUserMessage(t *testing.T, db *sql.DB, msg *common.Message) {
 		    unfurled_status_links
 		) VALUES (?,?,?,?,?,?,?,?,?,?,?,?)
 	`)
-	require.NoError(t, err)
+
+	s.Require().NoError(err)
 
 	links, err := json.Marshal(msg.UnfurledLinks)
-	require.NoError(t, err)
+	s.Require().NoError(err)
 
 	statusLinks, err := json.Marshal(msg.UnfurledStatusLinks)
-	require.NoError(t, err)
+	s.Require().NoError(err)
 
 	_, err = stmt.Exec(
 		msg.ID,
@@ -78,12 +90,12 @@ func createUserMessage(t *testing.T, db *sql.DB, msg *common.Message) {
 		links,
 		statusLinks,
 	)
-	require.NoError(t, err)
+	s.Require().NoError(err)
 }
 
-func httpGetReqRecorder(t *testing.T, handler http.HandlerFunc, reqURL string) *httptest.ResponseRecorder {
+func (s *HandlersSuite) httpGetReqRecorder(handler http.HandlerFunc, reqURL string) *httptest.ResponseRecorder {
 	req, err := http.NewRequest("GET", reqURL, nil)
-	require.NoError(t, err)
+	s.Require().NoError(err)
 
 	rr := httptest.NewRecorder()
 	handler.ServeHTTP(rr, req)
@@ -91,10 +103,8 @@ func httpGetReqRecorder(t *testing.T, handler http.HandlerFunc, reqURL string) *
 	return rr
 }
 
-func TestHandleLinkPreviewThumbnail(t *testing.T) {
-	db, logger := setupTest(t)
+func (s *HandlersSuite) TestHandleLinkPreviewThumbnail() {
 	previewURL := "https://github.com"
-
 	msg := common.Message{
 		ID: "1",
 		ChatMessage: &protobuf.ChatMessage{
@@ -109,31 +119,111 @@ func TestHandleLinkPreviewThumbnail(t *testing.T) {
 			},
 		},
 	}
-	createUserMessage(t, db, &msg)
+	s.createUserMessage(&msg)
 
 	// Test happy path.
 	reqURL := "/dummy?" + url.Values{"message-id": {msg.ID}, "url": {previewURL}}.Encode()
-	rr := httpGetReqRecorder(t, handleLinkPreviewThumbnail(db, logger), reqURL)
-	require.Equal(t, http.StatusOK, rr.Code)
-	require.Equal(t, msg.UnfurledLinks[0].ThumbnailPayload, rr.Body.Bytes())
-	require.Equal(t, "image/jpeg", rr.HeaderMap.Get("Content-Type"))
-	require.Equal(t, "no-store", rr.HeaderMap.Get("Cache-Control"))
+	rr := s.httpGetReqRecorder(handleLinkPreviewThumbnail(s.db, s.logger), reqURL)
+	s.Require().Equal(http.StatusOK, rr.Code)
+	s.Require().Equal(msg.UnfurledLinks[0].ThumbnailPayload, rr.Body.Bytes())
+	s.Require().Equal("image/jpeg", rr.HeaderMap.Get("Content-Type"))
+	s.Require().Equal("no-store", rr.HeaderMap.Get("Cache-Control"))
 
 	// Test bad requests.
 	reqURL = "/dummy?" + url.Values{"message-id": {msg.ID}}.Encode()
-	rr = httpGetReqRecorder(t, handleLinkPreviewThumbnail(db, logger), reqURL)
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Equal(t, "missing query parameter 'url'\n", rr.Body.String())
+	rr = s.httpGetReqRecorder(handleLinkPreviewThumbnail(s.db, s.logger), reqURL)
+	s.Require().Equal(http.StatusBadRequest, rr.Code)
+	s.Require().Equal("missing query parameter 'url'\n", rr.Body.String())
 
 	reqURL = "/dummy?" + url.Values{"url": {previewURL}}.Encode()
-	rr = httpGetReqRecorder(t, handleLinkPreviewThumbnail(db, logger), reqURL)
-	require.Equal(t, http.StatusBadRequest, rr.Code)
-	require.Equal(t, "missing query parameter 'message-id'\n", rr.Body.String())
+	rr = s.httpGetReqRecorder(handleLinkPreviewThumbnail(s.db, s.logger), reqURL)
+	s.Require().Equal(http.StatusBadRequest, rr.Code)
+	s.Require().Equal("missing query parameter 'message-id'\n", rr.Body.String())
 
 	// Test mime type not supported.
 	msg.UnfurledLinks[0].ThumbnailPayload = []byte("unsupported image")
-	createUserMessage(t, db, &msg)
+	s.createUserMessage(&msg)
 	reqURL = "/dummy?" + url.Values{"message-id": {msg.ID}, "url": {previewURL}}.Encode()
-	rr = httpGetReqRecorder(t, handleLinkPreviewThumbnail(db, logger), reqURL)
-	require.Equal(t, http.StatusNotImplemented, rr.Code)
+	rr = s.httpGetReqRecorder(handleLinkPreviewThumbnail(s.db, s.logger), reqURL)
+	s.Require().Equal(http.StatusNotImplemented, rr.Code)
+}
+
+func (s *HandlersSuite) TestHandleStatusLinkPreviewThumbnail() {
+	thumbnailPayload := []byte{0xff, 0xd8, 0xff, 0xdb, 0x0, 0x84, 0x0, 0x50, 0x37, 0x3c, 0x46, 0x3c, 0x32, 0x50}
+
+	contact := &protobuf.UnfurledStatusContactLink{
+		PublicKey: "PublicKey_1",
+		Icon: &protobuf.UnfurledLinkThumbnail{
+			Width:   10,
+			Height:  20,
+			Payload: thumbnailPayload,
+		},
+	}
+
+	unfurledContact := &protobuf.UnfurledStatusLink{
+		Url: "https://status.app/u/",
+		Payload: &protobuf.UnfurledStatusLink_Contact{
+			Contact: contact,
+		},
+	}
+
+	community := &protobuf.UnfurledStatusCommunityLink{
+		CommunityId: "CommunityId_1",
+		Icon: &protobuf.UnfurledLinkThumbnail{
+			Width:   30,
+			Height:  40,
+			Payload: thumbnailPayload,
+		},
+		Banner: &protobuf.UnfurledLinkThumbnail{
+			Width:   50,
+			Height:  60,
+			Payload: thumbnailPayload,
+		},
+	}
+
+	unfurledCommunity := &protobuf.UnfurledStatusLink{
+		Url: "https://status.app/c/",
+		Payload: &protobuf.UnfurledStatusLink_Community{
+			Community: community,
+		},
+	}
+
+	channel := &protobuf.UnfurledStatusChannelLink{
+		ChannelUuid: "ChannelUuid_1",
+		Community: &protobuf.UnfurledStatusCommunityLink{
+			CommunityId: "CommunityId_2",
+			Icon: &protobuf.UnfurledLinkThumbnail{
+				Width:   70,
+				Height:  80,
+				Payload: thumbnailPayload,
+			},
+			Banner: &protobuf.UnfurledLinkThumbnail{
+				Width:   90,
+				Height:  100,
+				Payload: thumbnailPayload,
+			},
+		},
+	}
+
+	unfurledChannel := &protobuf.UnfurledStatusLink{
+		Url: "https://status.app/cc/",
+		Payload: &protobuf.UnfurledStatusLink_Channel{
+			Channel: channel,
+		},
+	}
+
+	msg := common.Message{
+		ID: "1",
+		ChatMessage: &protobuf.ChatMessage{
+			UnfurledStatusLinks: &protobuf.UnfurledStatusLinks{
+				UnfurledStatusLinks: []*protobuf.UnfurledStatusLink{
+					unfurledContact,
+					unfurledCommunity,
+					unfurledChannel,
+				},
+			},
+		},
+	}
+
+	s.Require().NotNil(msg)
 }
