+++ server/server_test.go
@@ -119,11 +119,11 @@ func (s *ServerURLSuite) TestServer_MakeLinkPreviewThumbnailURL() {
 func (s *ServerURLSuite) TestServer_MakeStatusLinkPreviewThumbnailURL() {
 	s.Require().Equal(
 		baseURLWithCustomPort+"/status-link-preview/thumbnail?image-id=contact-icon&message-id=99&url=https%3A%2F%2Fstatus.app",
-		s.server.MakeStatusLinkPreviewThumbnailURL("99", "https://status.app", "contact-icon"))
+		s.server.MakeStatusLinkPreviewThumbnailURL("99", "https://status.app", ContactIcon))
 
 	s.testNoPort(
 		baseURLWithDefaultPort+"/status-link-preview/thumbnail?image-id=contact-icon&message-id=99&url=https%3A%2F%2Fstatus.app",
-		s.serverNoPort.MakeStatusLinkPreviewThumbnailURL("99", "https://status.app", "contact-icon"))
+		s.serverNoPort.MakeStatusLinkPreviewThumbnailURL("99", "https://status.app", ContactIcon))
 }
 
 func (s *ServerURLSuite) TestServer_MakeAudioURL() {
